Module x509::der::write

DER serialization APIs.

Functions

der_bit_string

Encodes an ASN.1 bit string using DER.

der_boolean

Encodes a boolean as an ASN.1 BOOLEAN using DER.

der_default

Serializes the given value if it is not equal to its default.

der_explicit

Wraps an ASN.1 data value in an EXPLICIT marker.

der_generalized_time

Encodes an ASN.1 GeneralizedTime using DER.

der_integer

Encodes a big-endian-encoded integer as an ASN.1 integer using DER.

der_integer_usize

Encodes a usize as an ASN.1 integer using DER.

der_null

Encodes an ASN.1 NULL using DER.

der_octet_string

Encodes an ASN.1 octet string using DER.

der_oid

Encodes an ASN.1 Object Identifier using DER.

der_sequence

Encodes the output of a sequence of serializers as an ASN.1 sequence using DER.

der_set

Encodes the output of a sequence of serializers as an ASN.1 set using DER.

der_utc_time

Encodes an ASN.1 UTCTime using DER.

der_utf8_string

Encodes an ASN.1 UTF8String using DER.